Use size_t for len

+++ src/input.h
@@ -2,7 +2,7 @@
 #define SCREEN_INPUT_H
 
 void  inp_setprompt (char *, char *);
-void  Input (char *, int, int, void (*)(char *, int, void *), char *, int);
+void  Input (char *, size_t, int, void (*)(char *, size_t, void *), char *, int);
 int   InInput (void);
 
 #endif /* SCREEN_INPUT_H */
